% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Makes the first word in Habitat lowercase}
\usage{
capitalize(ds)
}
\arguments{
\item{ds}{dataframe}
}
\value{
dataframe
}
\description{
Makes the first word in Habitat lowercase
}
