% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_abb.R
\name{dir_abb}
\alias{dir_abb}
\title{Directions abbreviated
changes the N E W S  to full words in the dataframe}
\usage{
dir_abb(df)
}
\arguments{
\item{df}{dataframe takes in a herbarium spreadsheet}
}
\value{
dataframe
}
\description{
Directions abbreviated
changes the N E W S  to full words in the dataframe
}
