% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.R
\name{lower}
\alias{lower}
\title{Makes the field notes column to lower case}
\usage{
lower(ds)
}
\arguments{
\item{ds}{dataframe - a ubc herbarium spreadsheet}
}
\value{
dataframe
}
\description{
Makes the field notes column to lower case
}
