% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{Mapping}
\usage{
mapping(ds, DMS = FALSE)
}
\arguments{
\item{ds}{dataframe - a herbarium spreadsheet}

\item{DMS}{logical - whether or not the coordinates are in degree minutes and seconds}
}
\value{
Web Map
}
\description{
Mapping
}
