% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oia.R
\name{oia}
\alias{oia}
\title{Originally Identified As}
\usage{
oia(g, q, s, sub, v, f, det, yr)
}
\arguments{
\item{g}{genus}

\item{q}{quantifier}

\item{s}{species}

\item{sub}{subspecies}

\item{v}{variety}

\item{f}{forma}

\item{det}{determined by}

\item{yr}{year}
}
\value{
Character
}
\description{
given the name and determiner and year and formats it into the Original ID as format
note must use group_by on a unique identifier when using mutate
}
\examples{
oia("gen","q","s","sub","var","for","det","yr")
}
